requirements.txt
django
djangorestframework
drf-spectacular
python-dotenv
pytest
pytest-django
requests
